// typescript/output/dynamic/types/tertiary.ts
// ==========================================
// Auto-generated file. Do not edit directly.
// ==========================================

// #region IMPORTS
import { Attachment, Collaborator, FieldSet } from "airtable";
import { RecordId } from "../../static/special-types";
// #endregion


// #region FIELD OPTIONS
// #endregion

// #region TERTIARY
    /** Field names for `Tertiary` */
export type TertiaryField = 
    "Name" |
    "Value"
    /** Field names for `Tertiary` */
export const TertiaryFields: TertiaryField[] = [
    "Name",
    "Value",
]

    /** Field IDs for `Tertiary` */
export type TertiaryFieldId = 
    "fldwzqKxsRnPZJ2Ll" |
    "fldjNLBh2UccM64h5"
    /** Field IDs for `Tertiary` */
export const TertiaryFieldIds: TertiaryFieldId[] = [
    "fldwzqKxsRnPZJ2Ll",
    "fldjNLBh2UccM64h5",
]

    /** Property names for `Tertiary` */
export type TertiaryFieldProperty = 
    "name" |
    "value"
    /** Property names for `Tertiary` */
export const TertiaryFieldPropertys: TertiaryFieldProperty[] = [
    "name",
    "value",
]

    /** Calculated fields for `Tertiary` */
export const TertiaryCalculatedFields: string[] = [
]
    /** Calculated fields for `Tertiary` */
export const TertiaryCalculatedFieldIds: string[] = [
]

    /** Writable fields for `Tertiary` */
export const TertiaryWritableFields: string[] = [
    "Name",
    "Value",
]
    /** Writable fields for `Tertiary` */
export const TertiaryWritableFieldIds: string[] = [
    "fldwzqKxsRnPZJ2Ll",
    "fldjNLBh2UccM64h5",
]
export const TertiaryFieldNameIdMapping: Record<TertiaryField, TertiaryFieldId> = {
    "Name": "fldwzqKxsRnPZJ2Ll",
    "Value": "fldjNLBh2UccM64h5",
}

export const TertiaryFieldIdNameMapping: Record<TertiaryFieldId, TertiaryField> = {
    "fldwzqKxsRnPZJ2Ll": "Name",
    "fldjNLBh2UccM64h5": "Value",
}

export const TertiaryFieldIdPropertyMapping: Record<TertiaryFieldId, TertiaryFieldProperty> = {
    "fldwzqKxsRnPZJ2Ll": "name",
    "fldjNLBh2UccM64h5": "value",
}

export const TertiaryFieldPropertyIdMapping: Record<TertiaryFieldProperty, TertiaryFieldId> = {
    "name": "fldwzqKxsRnPZJ2Ll",
    "value": "fldjNLBh2UccM64h5",
}

export const TertiaryFieldNamePropertyMapping: Record<TertiaryField, TertiaryFieldProperty> = {
    "Name": "name",
    "Value": "value",
}

export const TertiaryFieldPropertyNameMapping: Record<TertiaryFieldProperty, TertiaryField> = {
    "name": "Name",
    "value": "Value",
}

export interface TertiaryFieldSetIds extends FieldSet {
    //@ts-ignore
    fldwzqKxsRnPZJ2Ll?: string
    //@ts-ignore
    fldjNLBh2UccM64h5?: string
}

export interface TertiaryFieldSet extends FieldSet {
    //@ts-ignore
    "Name"?: string,
    //@ts-ignore
    "Value"?: string,
}

    /** View names for `Tertiary` */
export type TertiaryView = 
    "Grid view"
    /** View names for `Tertiary` */
export const TertiaryViews: TertiaryView[] = [
    "Grid view",
]

    /** View IDs for `Tertiary` */
export type TertiaryViewId = 
    "viwdp3tOB8ooOCvP4"
    /** View IDs for `Tertiary` */
export const TertiaryViewIds: TertiaryViewId[] = [
    "viwdp3tOB8ooOCvP4",
]

export const TertiaryViewNameIdMapping: Record<TertiaryView, TertiaryViewId> = {
    "Grid view": "viwdp3tOB8ooOCvP4",
}

export const TertiaryViewIdNameMapping: Record<TertiaryViewId, TertiaryView> = {
    "viwdp3tOB8ooOCvP4": "Grid view",
}

// #endregion